changed ints to doubles where needed int response files

+++ src/main/java/VisionApi/JSONClasses/Color.java
@@ -4,8 +4,8 @@ package VisionApi.JSONClasses;
  * Created by sierra on 4/22/16.
  */
 public class Color {
-    public int red;
-    public int green;
-    public int blue;
-    public int alpha;
+    public double red;
+    public double green;
+    public double blue;
+    public double alpha;
 }
